// spa/src/app/models/basket.ts
export interface Basket {
    id:      number;
    buyerId: string;
    items:   BasketItem[];
    PaymentIntentId?:string;
    clientSecret?:string;
}

export interface BasketItem {
    productId:  number;
    name:       string;
    price:      number;
    pictureUrl: string;
    brand:      string;
    type:       string;
    quantity:   number;
}